# swagger.yaml
swagger: "2.0"
info:
  title: dp-topic-api
  description: "This service is used to retrieve information about the Topic's used
  to categorize ONS data"
  version: 1.0.0
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
basePath: "/v1"
schemes:
  - http
tags:
  - name: "Public"
parameters:
  id:
    name: id
    description: "The ID of a topic"
    in: path
    type: string
    required: true
  content_options:
    description: "An option to select a subset of the possible types of content returned; can be one of: spotlight, articles, bulletins, methodologies, methodologyarticles, staticdatasets, timeseries, publications or datasets"
    name: "<type>"
    in: query
    required: false
    type: string
  lang:
    name: lang
    description: "the 2 character code of the language required in returned labels, e.g. cy for welsh"
    in: query
    type: string
    required: false
paths:
  /topics/{id}:
    get:
      tags:
        - "Public"
      summary: "Get a topic by ID"
      description: "provides a high level description of the topic and relevant links"
      parameters:
        - $ref: '#/parameters/id'
      produces:
        - "application/json"
      responses:
        200:
          description: "Json object containing information about the topic"
          schema:
            $ref: '#/definitions/Topic'
        404:
          description: "Topic not found"
        500:
          $ref: '#/responses/InternalError'
  /topics/{id}/subtopics:
    get:
      tags:
        - "Public"
      summary: "Get a list of subtopics"
      description: "Get a list of all documents for ID contained in the stored list of subtopics"
      parameters:
        - $ref: '#/parameters/id'
      produces:
        - "application/json"
      responses:
        200:
          description: "Json object containing an array of subtopics"
          schema:
            $ref: '#/definitions/ListOfTopics'
        404:
          description: "Topic not found"
        500:
          $ref: '#/responses/InternalError'
  /topics/{id}/content:
    get:
      tags:
        - "Public"
      summary: "Get a list of content"
      description: "Get a list of different content types of all documents for ID"
      parameters:
        - $ref: '#/parameters/id'
        - $ref: '#/parameters/content_options'
      produces:
        - "application/json"
      responses:
        200:
          description: "Json object containing an array(s) of content types"
          schema:
            $ref: '#/definitions/ListOfContent'
        400:
          description: "content query not recognised"
        404:
          description: "Content not found"
        500:
          $ref: '#/responses/InternalError'
  /navigation:
    get:
      tags:
        - "Public"
      summary: "Get a list for navigation"
      description: "Get a list of topics for site navigation"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/lang'
      responses:
        200:
          description: "Provides a hierarchical list of navigation items with their links, copy and localisation references"
          schema:
            $ref: '#/definitions/Navigation'
        500:
          $ref: '#/responses/InternalError'
responses:
  InternalError:
    description: "Failed to process the request due to an internal error"
  Unauthorised:
    description: "Failed to process the request due to being unauthorised"
definitions:
  Content:
    type: object
    properties:
      title:
        type: string
        description: "title of the information"
      type:
        type: string
        description: "Name of the containing type"
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/SelfHref'
          topic:
            $ref: '#/definitions/TopicLink'
      state:
        $ref: '#/definitions/State'
  Topic:
    type: object
    properties:
      description:
        type: string
        description: "A description of the topic"
      title:
        type: string
        description: "The title of the topic"
      keywords:
        type: array
        items:
          type: string
        description: "List of keywords that relate to the topic"
      state:
        $ref: '#/definitions/State'
      links:
        type: object
        properties:
          self:
            $ref: '#/definitions/SelfHref'
          subtopics:
            $ref: '#/definitions/SubtopicsLink'
          content:
            $ref: '#/definitions/ContentLink'
  SubtopicsLink:
    type: object
    description: "A link to the subtopics of this topic"
    properties:
      href:
        type: string
        description: "A URL to the subtopics of this topic"
  ContentLink:
    type: object
    description: "A link to the content for this topic"
    properties:
      href:
        type: string
        description: "A URL to the content for this topic"
  ListOfContent:
    type: object
    description: "A list of Content items"
    properties:
      count:
        $ref: '#/definitions/Count'
      offset:
        $ref: '#/definitions/Offset'
      limit:
        $ref: '#/definitions/Limit'
      total_count:
        $ref: '#/definitions/TotalCount'
      items:
        type: array
        items:
          $ref: '#/definitions/Content'
  ListOfTopics:
    type: object
    description: "A list of Topics"
    properties:
      count:
        $ref: '#/definitions/Count'
      offset:
        $ref: '#/definitions/Offset'
      limit:
        $ref: '#/definitions/Limit'
      total_count:
        $ref: '#/definitions/TotalCount'
      items:
        type: array
        items:
          $ref: '#/definitions/Topic'
  SelfHref:
    type: object
    properties:
      id:
        type: string
        description: "The id of the resource"
      href:
        type: string
        description: "The URL to the resource"
  State:
    description: |
      The state of the resource, can only be one of the following:
        * created
        * published
        * completed
    type: string
  TopicLink:
    type: object
    description: "A link to the topic for this content"
    properties:
      href:
        type: string
        description: "The URL to the topic"
      id:
        type: string
        description: "The id for the topic that this content is associated with"
  Count:
    type: integer
    description: "The number of results returned in this response"
  TotalCount:
    type: integer
    description: "The total number of results available"
  Limit:
    type: integer
    description: "The limit applied to the number of results returned"
  Offset:
    type: integer
    description: "The offset applied to the number of results returned"
  Description:
    type: string
    description: "A human-readable description as a string"
  Name:
    type: string
    description: "A hyphenated reference to the item"
  Label:
    type: string
    description: "The label to be displayed to a user, translated based on the lang parameter"
  Title:
    type: string
    description: "A string containing the title of something"
  Uri:
    type: string
    description: "The path to the resource on the web domain"
  Navigation:
    type: object
    description: "Resource describing the structure for the ONS website navbar."
    properties:
      description:
        type: string
      links:
        type: object
        properties:
          self:
            type: object
            properties:
              href:
                type: string
                description: "The URL to the resource"
      items:
        $ref: '#/definitions/ListOfNavigationItems'
  ListOfNavigationItems:
    type: array
    description: "A list of navigation items"
    items:
      type: object
      properties:
        description:
          $ref: '#/definitions/Description'
        label:
          $ref: '#/definitions/Label'
        links:
          type: object
          properties:
            self:
              $ref: '#/definitions/SelfHref'
        name:
          $ref: '#/definitions/Name'
        title:
          $ref: '#/definitions/Title'
        uri:
          $ref: '#/definitions/Uri'
        subtopics:
          type: array
          items:
            properties:
              description:
                $ref: '#/definitions/Description'
              label:
                $ref: '#/definitions/Label'
              links:
                type: object
                properties:
                  self:
                    $ref: '#/definitions/SelfHref'
              name:
                $ref: '#/definitions/Name'
              title:
                $ref: '#/definitions/Title'
              uri:
                $ref: '#/definitions/Uri'